% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a dataframe with two classes of the probability and cumulative ditribution
}
\description{
calculate the probability and cumulative distributions
}
\examples{
bin_cumulative(5, 0.5)
}
