% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a dataframe with two classes of the probability ditribution
}
\description{
calculate the distribution of probabilities of getting different number of successes
}
\examples{
bin_distribution(n = 5, p = 0.5)
}
