% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the kurtosis
}
\description{
calculate the kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
