% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the mean
}
\description{
calculate the mean
}
\examples{
bin_mean(10, 0.3)
}
