% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the mode
}
\description{
calculate the mode
}
\examples{
bin_mode(10, 0.3)
}
