% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(k, n, p)
}
\arguments{
\item{k}{number of successes}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the probability of getting k successes with probability p in n trials
}
\description{
calculate the probability of getting k successes with probability p in n trials
}
\examples{
bin_probability(k = 2, n = 5, p = 0.5)

bin_probability(k = 0:2, n = 5, p = 0.5)
}
