% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the skewness
}
\description{
calculate the skewness
}
\examples{
bin_skewness(10, 0.3)
}
