% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
a list with number of trials and probability of success
}
\description{
display the binomial random variable objects
}
\examples{
bin_variable(10, 0.3)
}
