% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(n, p)
}
\arguments{
\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
the variance
}
\description{
calculate the variance
}
\examples{
bin_variance(10, 0.3)
}
