% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/f2nr.R
\name{f2nr}
\alias{f2nr}
\title{Format to Number}
\usage{
f2nr(y)
}
\arguments{
\item{y}{A character, which is a number in format \code{"1.234,45"}}
}
\value{
A character, which is the number in format \code{"1234.45"}
}
\description{
This function reformats numbers, which are stored as characters,
with format 1.200,45 into 1200.45. The result is a character; use
\code{as.numeric()} to convert it to a number.

This can be necessary when trying to read text files that use
comma as decimal point and a dot as thousand separator for numbers.
}
\examples{
f2nr("1.234,45") # returns "1234.45"
f2nr("34,45") # returns "34.45"
}
\keyword{format}
\keyword{numeric,}

