% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{f_to_num}
\alias{f_to_num}
\title{Factor to Numeric Function}
\usage{
f_to_num(f)
}
\arguments{
\item{factor}{variable}
}
\description{
This function allows you to properly convert factor variables to numeric.
}
\examples{
f_to_num()
}
\keyword{factor}

