% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_groups}
\alias{generate_groups}
\title{Generate group indices according to a distribution}
\usage{
generate_groups(num_samples, num_groups, distribution = stats::runif,
  ...)
}
\arguments{
\item{num_samples}{natural number, number of draws from group labels}

\item{num_groups}{natural number, number of group labels}

\item{distribution}{distribition, e.g. runif, rnorm
Must accept one parameter. Results should be
non-negative.}

\item{...}{further parameters for the distribution}
}
\value{
\code{num_samples} samples of group indices
}
\description{
Generate group indices according to a distribution
}
\examples{
generate_groups(10, 2, stats::runif)
}
