% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_probabilities}
\alias{generate_probabilities}
\title{Generate relative probabilities according to a distribution}
\usage{
generate_probabilities(num_groups, distribution = stats::runif, ...)
}
\arguments{
\item{num_groups}{natural number, number of group labels}

\item{distribution}{distribition, e.g. runif, rnorm
Must accept one parameter. Results should be
non-negative.}

\item{...}{further parameters for the distribution}
}
\value{
\code{num_groups} samples from the distribution
}
\description{
The probabilities are intended to be used as weights in
functions like \code{sample}.
}
\examples{
generate_probabilities(10, stats::runif)
generate_probabilities(5, stats::rnorm, mean = 3, sd = 0.5)
generate_probabilities(6, function(n) { base::abs(stats::rnorm(n)) })
}
