% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{rank_outcomes}
\alias{rank_outcomes}
\title{Generate a tibble of ranked outcomes}
\usage{
rank_outcomes(outcomes)
}
\arguments{
\item{outcomes}{numeric}
}
\value{
a tibble in which \code{outcomes} (y) are sorted by rank (x)
}
\description{
Generate a tibble of ranked outcomes
}
\examples{
rank_outcomes(c(6, 5, 4))
}
