% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{replicate_df}
\alias{replicate_df}
\title{Repeat a dataframe-generating operation and collect results in a dataframe.}
\usage{
replicate_df(num_replications, df_generator, ...)
}
\arguments{
\item{num_replications}{natural number}

\item{df_generator}{function that generates a tibble}

\item{...}{additional parameters for df_generator}
}
\value{
a tibble of all rows of each generated tibble, with the replication
index labeled by a new variable, \code{replication}.
}
\description{
Repeat a dataframe-generating operation and collect results in a dataframe.
}
\examples{
replicate_df(3, function(n) {head(iris, n)}, 1)
replicate_df(3, function() { rank_outcomes(generate_probabilities(2)) })
}
