% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{stratify_by_group}
\alias{stratify_by_group}
\title{Stratify random values by group}
\usage{
stratify_by_group(num_groups, groups, distribution = stats::rnorm, ...)
}
\arguments{
\item{num_groups}{natural number, number of group values
to create}

\item{groups}{vector of natural numbers of at most
magnitude \code{num_groups}}

\item{distribution}{random value function that takes at
least one parameter, the number of values to return}

\item{...}{additional parameters for \code{distribution}}
}
\value{
a vector of random values of length
\code{length(groups)} with a functional dependence on
\code{groups}
}
\description{
Stratify random values by group
}
\examples{
num_groups <- 5
num_samples <- 20
groups <- generate_groups(num_samples, num_groups)
groups
v1 <- stratify_by_group(num_groups, groups)
v1
table(groups, v1)
v2 <- stratify_by_group(num_groups, groups, distribution = function(n) {base::sample.int(2, size = n, replace = TRUE)})
v2
table(groups, v2)
}
