% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_theme.R
\name{theme_modern}
\alias{theme_modern}
\title{Title}
\usage{
theme_modern(base_size = 11, base_family = "Lucida Sans Unicode",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size in points (11)}

\item{base_family}{base font family ('Lucida Sans Unicode')}

\item{base_line_size}{base line size in units of 22pt (\code{base_size/22})}

\item{base_rect_size}{base rectangle size in units of 22pt (\code{base_size/22})}
}
\value{
a ggtheme()-like object
}
\description{
Title
}
\examples{
library(extrafont)
windowsFonts()
extrafont::font_import()
extrafont::loadfonts(device="win")
windowsFonts()
ggplot2::ggplot(data = iris, ggplot2::aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  ggplot2::geom_point() +
  ggplot2::labs(title = 'Iris setosa is separable, but versicolor and verginica are not.') +
  theme_modern()
}
